- Changed the max playlist size to 2000 tracks - Added code to try to find the php.ini on unix (for dynamically loading extensions that are not compiled into the PHP executable) - Language files do not report to the log if missing (eg: en-gb) - System prefs do not update if the value in the DB is already set to the given value - run_background() specifies the PHP ini location if found - playlists for photos use a different URL on the PC to that used for the showcenter - Movie details are only looked for if there is an internet connection.

+++ branches/development/swisscenter/base/prefs.php
@@ -51,15 +51,21 @@
 
  function set_sys_pref( $name, $value)
  {
-   db_sqlcommand("delete from system_prefs where name='".strtoupper($name)."'");
-   $result = db_insert_row('system_prefs', array("NAME"=>strtoupper($name), "VALUE"=>$value) );
-
-   if (!$result)
-     send_to_log("Unable to store system preference '$name' = '$value'");
-   else
-     send_to_log("Set system preference '$name' to '$value'");
-
-   return $result;
+   // Only update if the value changes
+   if (db_value("select count(*) from system_prefs where name='".strtoupper($name)."' and value='$value'") == 0)
+   {
+     db_sqlcommand("delete from system_prefs where name='".strtoupper($name)."'");
+     $result = db_insert_row('system_prefs', array("NAME"=>strtoupper($name), "VALUE"=>$value) );
+  
+     if (!$result)
+       send_to_log("Unable to store system preference '$name' = '$value'");
+     else
+       send_to_log("Set system preference '$name' to '$value'");
+  
+     return $result;
+   }
+   else 
+     return true;
  }
  
 /**************************************************************************************************
